pyproject.toml
[project]
name = "ragbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "gdown>=5.2.1",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-mongodb>=0.11.0",
    "motor>=3.7.1",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "rapidocr-onnxruntime>=1.4.4",
    "uvicorn[standard]>=0.41.0",
]
